moved slightly forward with pairing code

+++ src/app/shared/reset.service.ts
@@ -3,10 +3,11 @@ import { HttpClient } from '@angular/common/http';
 
 @Injectable()
 export class ResetService {
-  resetMirror() {
-    this.http.delete('/api/picture/customer-photos').subscribe(res => { console.log(res);});
+  resetMirror(photos) {
+    if (photos) {
+      this.http.delete('/api/picture/customer-photos').subscribe(res => { console.log(res);});
+      }
     }
-
   constructor(private http: HttpClient) { }
 
 }
